pyproject.toml
[project]
name = "autotrainer"
version = "0.1.0"
description = "LLM-driven automated training system for PaddleFormers"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "textual>=0.50",
    "rich>=13.0",
    "pyyaml>=6.0",
    "filelock>=3.12",
    "psutil>=5.9",
    "pandas>=2.0",
    "huggingface_hub>=0.20",
    "matplotlib>=3.7",
    "openai>=1.0",
]

[project.optional-dependencies]
search = [
    "tavily-python>=0.3",
]
data = [
    "opendatalab>=0.0.1",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-textual>=0.2",
    "black>=23.0",
    "isort>=5.12",
    "ruff>=0.1",
]
all = [
    "autotrainer[search,data]",
]

[project.scripts]
autotrainer = "autotrainer.cli.main:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 119
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 119

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
